+++ frontend/src/components/molecules/SignupForm/styles.module.css
@@ -1,0 +1,38 @@
+.form {
+    width: 100%;
+    max-width: 600px;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    margin-top: 1rem;
+  }
+  
+  .field {
+    width: 100%;
+  }
+  
+  .label {
+    font-size: 0.875rem;
+    font-weight: bold;
+    color: #333;
+    display: block;
+  }
+  
+  .input {
+    width: 100%;
+    padding: 10px;
+    font-size: 1rem;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    margin-bottom: 12px;
+    box-sizing: border-box;
+  }
+  
+  .input:focus {
+    border-color: #007bff;
+    outline: none;
+  }
+  
+  .input::placeholder {
+    color: #888;
+  }
